+++ src/pages/FotoPage.tsx
@@ -1,0 +1,10 @@
+import { QuizLayout } from '@/features/quiz/components/QuizLayout'
+import { FotoUpload } from '@/features/quiz/components/FotoUpload'
+
+export function FotoPage() {
+  return (
+    <QuizLayout>
+      <FotoUpload />
+    </QuizLayout>
+  )
+}